Fix for loading selected session
Function `selectSession` in `MobileSessions` and `FixedSessions` calls
`sensors.tmpSelectedId()` assuming that `Sensors` has already fetched
data. It causes the problem when the response from `/api/sensors` comes
too late. The user would see the empty graph for selected session.

Loading and displaying single session for the user depends on many
factors and it should be refactored in the future.

But this should be preceded by a decision of how the UI will look like
(the view for display a single session), if we introduce pagination to
the list of all sessions.