fix: spin until UI is ready
Problem was that the spinner wasn't really spinning when it needed too.
Most of the times it was ending before the UI really was ready.

This is not a proper solution. In fact, it only spins while a http
request is in progress. But doesn't take into account the data
processing that happens afterwards. In any case, it seems like we don't
do much data processing cause the spinner disappears more or less when
the UI is ready. If that's not the case we could add a timeout when
setting the $isHttpInProgress flag back to false.

In short this seems good enough and saves us time