travis: fail build if deploy fails
In the after_success hook in case of failure the status of the build is
not affected. In other words, if the build is fine but the deploy fails,
the build stays green. That's bad cause developers wouldn't know if the
branch was deployed successfully or not.
This change fixes it.