feat(dropdowns)
The tab could be opened in three different ways:

* by opening /map (ie clicking on the maps link in the menu)
* by switching to fixed / mobile
* by copy / pasting a url (eg
/map#/map_fixed_sessions?map=%7B%22zoom%22:5,%22lat%22:37.09024,%22lng%22:-95.71289100000001,%22mapType%22:%22terrain%22%7D&data=%7B%22sensorId%22:%22Particulate%20Matter-AirBeam2-PM2.5%20(%C2%B5g%2Fm%C2%B3)%22,%22location%22:%7B%22address%22:%22%22,%22indoorOnly%22:false,%22streaming%22:true%7D,%22tags%22:%22%22,%22usernames%22:%22%22,%22time%22:%7B%22timeFrom%22:-60,%22timeTo%22:1379,%22dayFrom%22:1,%22dayTo%22:365,%22yearFrom%22:2017,%22yearTo%22:2018%7D,%22heat%22:%7B%22highest%22:150,%22high%22:55,%22mid%22:35,%22low%22:12,%22lowest%22:0%7D%7D&selectedSessionIds=%5B%5D&tmp=%7B%22selectedSensorId%22:%22%22%7D)

Before this PR the app always loaded as parameter / sensor the default
(ie particulate matter / ab2 pm 2.5) when in the url no sensorId was
specified or if sensorId was "". Unfortunately, all / all was equivalent
to sensorId === "" which loaded particulate matter / ab2 pm 2.5 and not
all / all. That happened for example if a user selected all / all and
then pasted the link to navigate back to the page.

Now the code persists sensorId === "all" when all / all is selected,
therefore pasting the link works.

Also, on the first load the sensors weren't filtered by parameter when the page was loaded.
In fact, the default selection was in place particulate matter / ab2 pm
2.5 but the sensors in the dropdown contained all of them (eg humidity,
sound level)