feat(usernames filter): do not show if only indoor checked
When only indoor is checked users should not be able to filter sessions
by usernames.
Since we are going to show "anonymous" as usernames for indoor sessions,
if users were able to filter by username they could try to guess to whom
a session belongs.
Users could still sniff from the network requests that info but the
client is fine with that for now.