feat(new-markers): set map markers to empty array only when removing markers
In order to remove a marker from google maps one has to set marker.setMap(null) - this way the marker dissapears from the view. But it's still in the array where we collect all markers. So at the same time we have to set map.markers = [].
Before this PR it was done all over the place and it was causing bugs. Cause if you set map.markers = [] before you delete this markers from the view, they stay visible and they will not be removed by map.removeAllMarkers, as it's looping over this array.