feat(fixed-clustering): use smaller gridResolution so the rectangles are big enough to contain all sessions
The clusterer select sessions that should be clustered. It returns them as an array and doesn't say which session should belong to which cluster.
In a separate part of code the sessions are grouped into rectangles based on location. If the rectangles are too small it might happen that each session will end up with a separate rectangle. In order for neighboring sessions to be in the same rectangle the rectangles should be biger. Hence the reason for smaller resolution.