fix(session#as_json): duplicated streams in json
In Rails v3.2 `as_json` was serializing streams as `:streams => []` then
we were overriding that with `map_of_streams` (ie hash). In v4.2 Rails serializes
`"streams" => []` and the override actually added `:streams => []`. That
way we had two `"streams":` in the final JSON and the mobile app was
failing the sync.