feat(time-filter): filter sessions based on time ranges
As discussed with the client we want to include all sessions that somehow overlap with the selected time range, not just the ones that fully fit inside the range.

This might cause a bug for fixed session, since it's possible to stop and restart the fixed session. For example say a fixed session was started and stopped on 1.01  and then restarted on the 5.01. If a user searches for sessions between 2.01 - 3.01 this session will be returned even though it doesn't contain any measurements for this time. This was discussed with the, we decided include more session is acceptable. It's much worse if we miss out some sessions.