fix(heat): updating on input
Problem was that when changing the maximum the slider would update on
every input. If the user had the handles set at 20, 40, 60 and they
started typing 100 in the max number field, the updates of the max
would have been 1 then 10 then 100. On the first update all handles
would have been set to 1 which is a problem.
Changing the field to text instead of number is done because in Chrome
only the input event is triggered pressing on the arrows
(806bc67813bf9aaa251785f7a0b1815acb5c5697).