fix(heatmap): set extremes first
Setting the extremes after the middle values was creating a bug. In
fact, when resetting with thresholds1 = { value = 15, default = 0 } and
thresholds2 = { value = 15, default = 12 }, only thresholds1 would be
reset to { value = 0, default = 0 }. That's because first the code would
try to reset thresholds2 but that would be a noop because it cannot set
the first handle lower than the minimum.