fix(tooltip): use one tooltip instance
before this change there was a bug when user opened a tooltip window and clicked again on the button to open a tooltip window. That second click would try to create a second tooltip instance using the same id as the first instance. That would return a null instead of creating a instance. The SetContent throw an error when executed on null.