ui: add back outlines to tabbable elements
Rationale: it's not a good practice to remove visual indicators of
navigation, and if we do so we need to provide an alternative. And I
don't want to do that just for some elements and not others because it
won't look good anyway. So I'd leave default outlines for all tabbable
elements and at some point, in a future iteration, work on nicer
indicators for all elements.