feat(graph-tooltip): switch tooltip position so it doesn't overlap with the polt line
Before this change tooltip was always displayed above the measurement point. This meant that at the top of the graph it was overlaping with the plot line.
This change assures that the tooltip swithes positions, it is located belove the measurement points in the top part of the graph and above the measurements in the bottom and middle part of the graph.