fix: Put filters into a div instead of a form
It solves a fascinating bug: submitting location was changing the state of "Redo search when map is
moved". This is because pressing enter on a form acts as pressing the
neares button of type submit (which is the default type). So pressing
enter while using filters was 'clicking' the refresh time range button,
which in turn was executing fetch sessions. And receiving new sessions
list changes "Redo Search in Map" back into "Redo search when map is
moved".

It's worth noting that the fetched sessions where incorrect. The command
to fetch session was triggered before the map had time to move to the
new location, so the sessions were for the old location.