fix(crowd map): group by query
Measurement has default_scope { order('time ASC') } which influences
every query on the model. We should move away from default scopes cause
they create this sort of problems. In other words, queries should be
explicit, the order should be used where necessary not in a
default_scope. This is just patching the problem for now