fix(tags): duplicate tag_list in json
Commit c64ed5a broke the code. It seems like acts-as-taggable-on
from version 6 started adding tag_list in the json (as_json). Since
in the json the key tag_list is a string, when merging :tag_list => ...,
the hash had the key twice (once as a string and once as a symbol). In
the end the json returned had twice that key as a string.