add a critical queue to process session stopped email alerts
these jobs have to be performed every half an hour so they can't be
added to the end of default queue (which might be long sometimes).
Ideally I would like to process every job in critical queue before any
job in other queues, but that is impossible to achieve with sidekiq as
far as I know. So the critical queue has the same weight as the default,
but since it's going to have a very small number of jobs this actually
guarantees that the jobs will be performed instantanenosly.