Fix map container height issues
The trick is to set overflow: hidden on the parent element containing filters and the map when filters are collapsed. Then we don't get the extra white space below the map, resulting from the fact that filters take up more vertical space, and we're still able to scroll the filters, because overflow is on when they're expanded.