fix(crowdmap)
The URL kept growing because `data` was nested one level deeper every
time the resolution was clicked. It started with { data: {} } then {
data: { data: {} } } and so on. That created a problem with the link
shortener. After some level of nesting it stopped working and returned
the full url every time. Notice that the app was working because the
only the crowdmap uses `updateData` and the updated value was at the
right nesting level. It was the old values being nested each time.