fix: fixed clusters redrawn when closing an info window
To cluster markers we use markerclustererplus:
https://github.com/googlemaps/v3-utility-library/tree/8cbe1dc36ff7ee6eb9ec199b9e38cb7dd7bde6f8/packages/markerclustererplus
As you can see from the examples, zooming in or out makes the library
redraw all the clusters.
Whenever performing a `fitBounds` GoogleMaps adds a bit of margin around
the passed bounds. That resulted in `_fitBoundsWithoutPanOrZoomCallback`
zooming out when calling `fitBounds` and zooming back in with `setZoom`.
The change of zoom was so fast the map did not even perform it. However,
markerclustererplus redrew clusters after closing an info window because of
the "closeclick" callback calling `fitBounds`.