fix(openaq): round to 9 decimal digits
The measurements that we receive from OpenAq are each in the following format

{"date":{"utc":"2020-04-29T04:00:00.000Z","local":"2020-04-29T06:00:00+02:00"},"parameter":"pm25","value":4.6,"unit":"¬µg/m¬≥","averagingPeriod":{"unit":"hours","value":1},"location":"FR34038","city":"Cher","country":"FR","coordinates":{"latitude":46.8148188300394,"longitude":2.61008200746779},"attribution":[{"name":"EEA","url":"http://www.eea.europa.eu/themes/air/air-quality"}],"sourceName":"EEA France","sourceType":"government","mobile":false}

Sometimes latitude or longitude have a different amount of decimal digits than the ones we end up storing in the database. For example, for the measurement above latitude is "46.8148188300394" but we store "46.81481883".

Whenever we get a new measurement from OpenAq we either append it to a session and stream if those exist or create a new session and stream if it's a measurement coming from a new OpenAq station. Here "new" means we have not imported a measurement coming from that station yet.

OpenAq does not provide any unique identifiers for each station. Thus, we identify them by the "latitude, longitude, stream name" triplet. In other words, two measurements with the same "latitude, longitude, stream name" are considered coming from the same OpenAq station.

The code does something like this:

if can_find_session_and_stream_by_triplet
then append_measurement_to_session_and_stream
else create_session and create_stream and append_measurement_to_session_and_stream

The problem lies in "can_find_session_and_stream_by_triplet". In fact, using the same example as above, the code tries to find the session and stream that are at latitude "46.8148188300394". Since in the database the session and the stream for that measurement have latitude "46.81481883", they are never found.

Therefore, every station that reports coordinates with a bigger amount of decimal digits than the one we store in the database ends up creating a new session and stream instead of appending to the already persisted session and stream for that station.