feat: (mobile)update measurements legend design on calendar
Using mobile Map Legend design on calendar with additional request to make the color boxes equal length (as this design is static - no draggable sliders - there seems to be no point for them to be of specifiic lengths).

We can change the draggability post MVP, but it will require quite a lot of mathematical finesse to calculate proportions within the grid, and I think it's not worth it atm.

related trello ticket: https://trello.com/c/UJ6I9HEA/1904-mobile-devices-fixed-mobile-sessions-measurements-legend-is-obscured